pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
Flask = "^2.0.2"
azure-batch = "^12.0.0"
azure-storage-blob = "^12.9.0"
azure-storage-file-share = "^12.6.0"
config = "^0.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"